Update Session.php
setStepData & getStepData:
According to the function code, $step must not be an array as previously stated. It has to be either string or int. Since strings are used to identify checkout steps (like "billing"), I assumed this to be the correct type.

setStepData only:
The function does not perform any type check on $value, thus it can be of any type. Leaving out $value and calling the function with an array [data => value] is also possible and now reflected in the phpdoc comments.

getStepData only:
$step can be null and will return all steps' data (multi-dimensional array) or all of the given step's data (array) or the defined data from the given step (mixed), false otherwise.
I know that "array" is part of "mixed", but since it seems like a different context, i mentioned that additionally.